Update README, workspace.xml, and add pre-commit config
These changes primarily involve revisions to the README and workspace.xml files. Additional instructions on running tests without Docker were added in the README. Changes were made to the workspace.xml file, notable ones being changes in task summaries and counters, as well as adjustments to the duration of work items. A pre-commit config for the Black code formatter was added.